organize public mysql files
